pyproject.toml
[project]
name = "airpot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[all]>=0.124.4",
    "icecream>=2.1.8",
    "langchain>=1.2.0",
    "langchain-core>=1.2.1",
    "langchain-openai>=1.1.3",
    "langgraph>=1.0.5",
    "langgraph-checkpoint-postgres>=3.0.2",
    "psycopg-pool>=3.3.0",
    "psycopg[binary]>=3.3.2",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "redis>=6.4.0",
    "pyyaml>=6.0.3",
    "streamlit>=1.52.1",
    "telethon>=1.42.0",
    "uvicorn>=0.38.0",
]
